Function rustunnel::sandbox::close_all_fds

pub fn close_all_fds(keep_fds: &BTreeSet<RawFd>) -> Result<(), Error>

Closes all file descriptors except for those in keep_fds.

A sandboxed process is allowed to read and write to open file descriptors, but not to open new file descriptors. Therefore, the sandbox will be effective only if the sandboxed process has only harmless file descriptors open (stdin, stdout, and stderr, for example). This function is especially useful after a process fork() (like when being spawned by Command::new, for example) since any file descriptors open in the parent process may still be open in the child.

Portability

Currently this function only supports Linux with a mounted /proc, and fails otherwise.

Errors

If any error occurred closing any file descriptors, then an error is returned.